Dependencies: Added support of `numpy` v2

+++ src/silx/math/test/test_histogramnd_vs_np.py
@@ -1,5 +1,5 @@
 # /*##########################################################################
-# Copyright (C) 2016-2021 European Synchrotron Radiation Facility
+# Copyright (C) 2016-2024 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -608,7 +608,7 @@ class _TestHistogramnd(unittest.TestCase):
         sample = np.repeat(self.sample, n_repeat)
         n_bins = int(1e6)
         result_c = histogramnd(sample, self.histo_range, n_bins, last_bin_closed=True)
-        result_np = np.histogramdd(sample, n_bins, range=self.histo_range)
+        result_np = np.histogramdd(sample, n_bins, range=self.histo_range.astype(np.float64))
         for i_edges, edges in enumerate(result_c[2]):
             self.assertTrue(
                 np.allclose(edges, result_np[1][i_edges]),
